requirements.txt
Flask==1.1.2
flask_marshmallow==0.14.0
Flask_SQLAlchemy==2.4.3
PyJWT==2.3.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
gunicorn